# docs/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: alascup-agent API
  version: 0.1.0
  description: web-server OpenAPI 规范，定义前端与后端的 REST + SSE 契约。
servers:
  - url: /api
paths:
...
  /tool-requests/{request_id}/approval:
    post:
      summary: 审批工具请求
      description: 前端对 tool_approval_required 事件中的 request_id 做出批准或拒绝决策。后端通过内存 ApprovalBridge 唤醒原 SSE 连接中的 AgentLoop；过期由服务端内部超时产生，不能由客户端提交。
      parameters:
        - in: path
          name: request_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolApproval"
      responses:
        "200":
          description: 审批已接收。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolApprovalResult"
        "400":
          description: approval_status 非法；请求值必须为 APPROVED 或 REJECTED。
        "404":
          description: 请求不存在、已处理或已过期。
components:
  schemas:
...
    ToolApproval:
      type: object
      required: [approval_status]
      additionalProperties: false
      properties:
        approval_status:
          type: string
          enum: [APPROVED, REJECTED]
        reason:
          type: string
          nullable: true
    ToolApprovalResult:
      type: object
      required: [request_id, approval_status]
      description: Approval API response for accepted client decisions. Timeout is represented by 404 or server-side lifecycle state, not by this response body.
      properties:
        request_id:
          type: string
          format: uuid
        approval_status:
          type: string
          enum: [APPROVED, REJECTED]
        reason:
          type: string
          nullable: true
...